+++ packages/loader/src/index.ts
@@ -1,7 +1,12 @@
 import type { Sandbox } from '@qiankunjs/sandbox';
 import { qiankunHeadTagName } from '@qiankunjs/sandbox';
-import type { BaseTranspilerOpts } from '@qiankunjs/shared';
-import { Deferred, moduleResolver as defaultModuleResolver, QiankunError, transpileAssets } from '@qiankunjs/shared';
+import type {
+  AssetsTranspilerOpts,
+  BaseTranspilerOpts,
+  NodeTransformer,
+  ScriptTranspilerOpts,
+} from '@qiankunjs/shared';
+import { Deferred, prepareDeferredQueue, QiankunError } from '@qiankunjs/shared';
 import { createTagTransformStream } from './TagTransformStream';
 import { isUrlHasOwnProtocol } from './utils';
 import WritableDOMStream from './writable-dom';
@@ -19,8 +24,18 @@ type Entry = HTMLEntry;
 //
 export type LoaderOpts = {
   streamTransformer?: () => TransformStream<string, string>;
-  nodeTransformer?: typeof transpileAssets;
-} & BaseTranspilerOpts & { sandbox?: Sandbox };
+  nodeTransformer?: NodeTransformer;
+} & Omit<BaseTranspilerOpts, 'moduleResolver'> & { sandbox?: Sandbox };
+
+const isExternalScript = (script: HTMLScriptElement): boolean => {
+  return script.tagName === 'SCRIPT' && !!(script.src || script.dataset.src);
+};
+const isEntryScript = (script: HTMLScriptElement): boolean => {
+  return isExternalScript(script) && script.hasAttribute('entry');
+};
+const isDeferScript = (script: HTMLScriptElement): boolean => {
+  return isExternalScript(script) && script.hasAttribute('defer');
+};
 
 /**
  * @param entry
@@ -28,22 +43,12 @@ export type LoaderOpts = {
  * @param opts
  */
 export async function loadEntry<T>(entry: Entry, container: HTMLElement, opts: LoaderOpts): Promise<T | void> {
-  const {
-    fetch,
-    nodeTransformer = transpileAssets,
-    streamTransformer,
-    sandbox,
-    moduleResolver = (url: string) => {
-      return defaultModuleResolver(url, container, document.head);
-    },
-  } = opts;
+  const { fetch, streamTransformer, sandbox, nodeTransformer } = opts;
 
   const res = isUrlHasOwnProtocol(entry) ? await fetch(entry) : new Response(entry, { status: 200, statusText: 'OK' });
   if (res.body) {
+    let foundEntryScript = false;
     const entryScriptLoadedDeferred = new Deferred<T | void>();
-    const isEntryScript = (script: HTMLScriptElement): boolean => {
-      return script.hasAttribute('entry');
-    };
     const onEntryLoaded = () => {
       // the latest set prop is the entry script exposed global variable
       if (sandbox?.latestSetProp) {
@@ -54,6 +59,11 @@ export async function loadEntry<T>(entry: Entry, container: HTMLElement, opts: L
         entryScriptLoadedDeferred.resolve({} as T);
       }
     };
+
+    // defer scripts must wait until the entry html loaded
+    const deferQueue: Array<Deferred<void>> = [];
+    const { deferred: entryHTMLLoadedDeferred, queue: queueEntryHTMLDeferred } = prepareDeferredQueue(deferQueue);
+    queueEntryHTMLDeferred();
 
     let readableStream = res.body.pipeThrough(new TextDecoderStream());
 
@@ -76,21 +86,47 @@ export async function loadEntry<T>(entry: Entry, container: HTMLElement, opts: L
       )
       .pipeTo(
         new WritableDOMStream(container, null, (clone, node) => {
-          const transformedNode = nodeTransformer(clone, entry, {
+          let transformerOpts: AssetsTranspilerOpts = {
             fetch,
             sandbox,
-            moduleResolver,
             rawNode: node as unknown as Node,
-          });
+          };
+
+          let queueDeferScript: () => void;
+          const deferScriptMode = isDeferScript(node as unknown as HTMLScriptElement);
+          if (deferScriptMode) {
+            const { deferred, prevDeferred, queue } = prepareDeferredQueue(deferQueue);
+            transformerOpts = {
+              ...transformerOpts,
+              scriptTranspiledDeferred: deferred,
+              prevScriptTranspiledDeferred: prevDeferred,
+            } as ScriptTranspilerOpts;
+            queueDeferScript = queue;
+          }
+
+          const transformedNode = nodeTransformer ? nodeTransformer(clone, transformerOpts) : clone;
 
           const script = transformedNode as unknown as HTMLScriptElement;
+
+          // the script have no src attribute after transpile, indicating that the script needs to wait for the src to be filled
+          if (deferScriptMode && !script.hasAttribute('src')) {
+            queueDeferScript!();
+          }
 
           /*
            * If the entry script is executed, we can complete the entry process in advance
            * otherwise we need to wait until the last script is executed.
            * Notice that we only support external script as entry script thus we could do resolve the promise after the script is loaded.
            */
-          if (script.tagName === 'SCRIPT' && (script.src || script.dataset.src) && isEntryScript(script)) {
+          if (isEntryScript(script)) {
+            if (foundEntryScript) {
+              throw new QiankunError(
+                `You should not set multiply entry script in one entry html, but ${entry} has at least 2 entry scripts`,
+              );
+            }
+
+            foundEntryScript = true;
+
             const onScriptComplete = (
               prevListener: typeof HTMLScriptElement.prototype.onload | typeof HTMLScriptElement.prototype.onerror,
               event: Event,
@@ -103,7 +139,7 @@ export async function loadEntry<T>(entry: Entry, container: HTMLElement, opts: L
                   onEntryLoaded();
                 } else {
                   entryScriptLoadedDeferred.reject(
-                    new QiankunError(`entry ${entry} load failed as entry script ${script.dataset.src} load failed}`),
+                    new QiankunError(`entry ${entry} load failed as entry script ${script.src} load failed}`),
                   );
                 }
               }
@@ -119,14 +155,17 @@ export async function loadEntry<T>(entry: Entry, container: HTMLElement, opts: L
         }),
       )
       .then(() => {
-        // while the entry html stream is finished but there is no entry script found(entryScriptLoadedDeferred is not be resolved)
-        // we could use the latest set prop in sandbox to resolve the entry promise
-        if (!entryScriptLoadedDeferred.isSettled()) {
+        // while the entry html stream is finished but there is no entry script found
+        // we could use the latest set prop in sandbox to resolve the entry promise as fallback
+        if (!foundEntryScript) {
           onEntryLoaded();
         }
+
+        entryHTMLLoadedDeferred.resolve();
       })
       .catch((e) => {
         entryScriptLoadedDeferred.reject(e);
+        entryHTMLLoadedDeferred.reject(e);
       });
 
     return entryScriptLoadedDeferred.promise;