+++ packages/sandbox/src/patchers/dynamicAppend/types.ts
@@ -2,12 +2,14 @@
  * @author Kuitos
  * @since 2023-05-04
  */
-import type { BaseLoaderOpts, transpileAssets } from '@qiankunjs/shared';
+import type { BaseLoaderOpts, NodeTransformer } from '@qiankunjs/shared';
+import type { Deferred } from '@qiankunjs/shared';
 import type { Sandbox } from '../../core/sandbox';
 
 export type SandboxConfig = {
   appName: string;
   sandbox: Sandbox;
   dynamicStyleSheetElements: Array<HTMLStyleElement | HTMLLinkElement>;
-  dynamicExternalSyncScriptElements: HTMLScriptElement[];
-} & BaseLoaderOpts & { nodeTransformer?: typeof transpileAssets };
+  dynamicExternalSyncScriptDeferredList: Array<Deferred<void>>;
+  nodeTransformer: NodeTransformer;
+} & BaseLoaderOpts;