+++ packages/shared/src/assets-transpilers/types.ts
@@ -15,6 +15,8 @@ export type BaseTranspilerOpts = BaseLoaderOpts & {
 
 export type AssetsTranspilerOpts = BaseTranspilerOpts & { rawNode: Node };
 
+export type NodeTransformer = <T extends Node>(node: T, opts: Omit<AssetsTranspilerOpts, 'moduleResolver'>) => T;
+
 export type ScriptTranspilerOpts = AssetsTranspilerOpts &
   (
     | { prevScriptTranspiledDeferred: Deferred<void>; scriptTranspiledDeferred: Deferred<void> }