+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @qiankunjs/sandbox
 
+## 0.0.1-rc.12
+
+### Patch Changes
+
+- bd12dbad: fix: defer scripts should wait until html loaded
+- Updated dependencies [bd12dbad]
+  - @qiankunjs/shared@0.0.1-rc.9
+
+## 0.0.1-rc.11
+
+### Patch Changes
+
+- 98b071bf: feat: support defer scripts and keep the executing order to consist with browser
+- Updated dependencies [98b071bf]
+  - @qiankunjs/shared@0.0.1-rc.8
+
+## 0.0.1-rc.10
+
+### Patch Changes
+
+- f2af2e36: feat: extract NodeTransformer type to shared package
+- Updated dependencies [f2af2e36]
+  - @qiankunjs/shared@0.0.1-rc.7
+
+## 0.0.1-rc.9
+
+### Patch Changes
+
+- d3e9872d: feat(sandbox): use cloneNode api instead of importNode for compatible
+- 7cc06bd4: feat(loader): add lru cache for assets fetch by default
+- Updated dependencies [54b0878e]
+- Updated dependencies [7ba95cf2]
+- Updated dependencies [312abbc7]
+- Updated dependencies [6f074136]
+  - @qiankunjs/shared@0.0.1-rc.6
+
 ## 0.0.1-rc.8
 
 ### Patch Changes