+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @qiankunjs/shared
 
+## 0.0.1-rc.9
+
+### Patch Changes
+
+- bd12dbad: fix: defer scripts should wait until html loaded
+
+## 0.0.1-rc.8
+
+### Patch Changes
+
+- 98b071bf: feat: support defer scripts and keep the executing order to consist with browser
+
+## 0.0.1-rc.7
+
+### Patch Changes
+
+- f2af2e36: feat: extract NodeTransformer type to shared package
+
+## 0.0.1-rc.6
+
+### Patch Changes
+
+- 54b0878e: feat(loader): compatible with defer entry script
+- 7ba95cf2: feat: change script src before it execute thus we can be more consistent with the native browser logic
+- 312abbc7: feat: remove lru-cache and move wrapFetch to shared package
+- 6f074136: feat(transpiler): assets transpiler should work well while sandbox disabled
+
 ## 0.0.1-rc.5
 
 ### Patch Changes